test/web-platform-tests/to-upstream/jsdom-only/failure-example-2.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!-- This test exists to show off the to-upstream-expectations.yaml file. -->

<script>
"use strict";

test(() => {
  assert_true(false);
}, "failing test");

test(() => {
  assert_true(true);
}, "passing test");
</script>
